chore: export type and content
