+++ frontend/src/types/sidebar.ts
@@ -4,11 +4,17 @@ export enum SidebarPages {
 	HEIRS = 'heirs',
 	INHERITANCE = 'inheritance',
 	SUCCESSION = 'succession',
-	FINAL = 'final'
+	FINAL = 'final',
 }
 
 export enum SidebarButtonState {
 	ACTIVE = 'active',
 	DEFAULT = 'default',
 	DISABLED = 'disabled',
+}
+
+export type SidebarElement = {
+	page: SidebarPages
+	title: string
+	description?: string
 }
